some other small changes

+++ include/JGadget/hashcode.h
@@ -4,5 +4,5 @@
 
 namespace JGadget
 {
-    u16 getHashCode(const char *);
+    u32 getHashCode(const char *);
 };
